+++ CHANGELOG.md
@@ -54,6 +54,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   instrumentation using the [`tracing`](https://tracing.rs/) ecosystem.
 - Moved all `database()` functions to `ServerConnection`. This allows fully
   generic code to be written against a "server".
+- Added `listen_for_shutdown()` which listens for SIGINT and SIGQUIT and attemps
+  to shut the server down gracefully.
 
 ### Changed